refactor: Make `get_case_strategy` dependent on the used spec
It will unify interfaces for creating Hypothesis strategies
