+++ src/schemathesis/specs/openapi/constants.py
@@ -1,0 +1,8 @@
+LOCATION_TO_CONTAINER = {
+    "path": "path_parameters",
+    "query": "query",
+    "header": "headers",
+    "cookie": "cookies",
+    "body": "body",
+    "formData": "form_data",
+}
